Bump dependencies and version
Bump the pinned dependencies and bump the package version since the
google-cloud-pubsub bump will make it (in)compatible with other Google
packages.

Remove the pinning of dev-dependencies since it is not that necessary
for a library (and they are very seldom bumped anyway).

Change the dependency of comet-core to make releases easier.